MONITOR/requirements.txt
python-dotenv
requests
telethon